Support user's settings for sending checkins to Twitter and Facebook.
Keep an object with user's data as propery of fc object.
Get user's setting from 4sq and append them to user object and update localStorage.
When checking in, get settings for sending to twitter and facebook and format the request.
I changed the way the user is kept in local storage.'