Implemented venue search by name and adding venue.
Added search form after the venue list, so that the user can search venues by name if the venue he wants to
check in is not in the list. Also implemented adding venues with pre-filling the form with address data
either from position.address (Firefox extension to GeoLocation position object) or calling Google geocode API
and constructing address object identical to position.address.